pyscripts/requirements.txt
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
click==8.1.6
colorzero==1.1
consonance==0.1.5
cryptography==41.0.4
dbus-python==1.2.16
dissononce==0.34.3
distro==1.5.0
EasyProcess==1.1
emoji==2.8.0
entrypoint2==1.1
exceptiongroup==1.1.3
Flask==2.3.2
get-mac==0.9.2
gpiozero==1.6.2
gunicorn==21.2.0
gyp==0.1
h11==0.14.0
idna==2.10
importlib-metadata==6.8.0
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
mysql-connector-python==8.1.0
numpy==1.24.0
outcome==1.3.0
packaging==23.1
pandas==1.5.2
picamera2==0.3.12
pidng==4.0.9
piexif==1.1.3
Pillow==10.1.0
protobuf==4.21.12
pushbullet.py==0.12.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==1.10.13
pyfiglet==1.0.2
PyGetWindow==0.0.9
PyGObject==3.38.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-apt==2.2.1
python-axolotl==0.2.2
python-axolotl-curve25519==0.4.1.post2
python-dateutil==2.8.2
python-magic==0.4.27
python-prctl==1.7
python-snap7==1.2
python3-xlib==0.15
pytweening==1.0.7
pytz==2022.7
PyYAML==6.0.1
requests==2.25.1
RPi.GPIO==0.7.0
scapy==2.5.0
selenium==4.14.0
setproctitle==1.3.2
simplejpeg==1.6.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spidev==3.5
ssh-import-id==5.10
toml==0.10.1
transitions==0.9.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==1.26.5
v4l2-python3==0.3.2
verlat==0.1.0.post1
wappdriver==0.6.6
websocket-client==1.6.4
Werkzeug==2.3.6
wikipedia==1.4.0
wsproto==1.2.0
XlsxWriter==3.0.3
yowsup==3.3.0
zipp==3.16.2
